Clases principales terminadas (P1). Falta programa principal.

+++ src/deepspace/WeaponType.java
@@ -7,13 +7,14 @@ package deepspace;
 /**
  *
  * @author antoniojrr
+ * @author kessler
  */
 
 
 public enum WeaponType { 
-        LASER(2.0),
-        MISSILE(3.0),
-        PLASMA(4.0);
+        LASER((float) 2.0),
+        MISSILE((float) 3.0),
+        PLASMA((float) 4.0);
         
         private final float power;
         WeaponType(float power) { this.power=power; }
